# backend/openapi.json
{
  "components": {
    "schemas": {
      "AccountBalance": {
        "properties": {
          "balance_minor": {
            "title": "Balance Minor",
            "type": "integer"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "currency",
          "balance_minor"
        ],
        "title": "AccountBalance",
        "type": "object"
      },
      "AdminOverview": {
        "properties": {
          "kill_switches": {
            "items": {
              "$ref": "#/components/schemas/KillView"
            },
            "title": "Kill Switches",
            "type": "array"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/MemberRow"
            },
            "title": "Members",
            "type": "array"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunView"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "members",
          "runs",
          "kill_switches"
        ],
        "title": "AdminOverview",
        "type": "object"
      },
      "ApprovalRequest": {
        "additionalProperties": false,
        "properties": {
          "content_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Content Hash",
            "type": "string"
          },
          "decision": {
            "enum": [
              "APPROVE",
              "REJECT"
            ],
            "title": "Decision",
            "type": "string"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          }
        },
        "required": [
          "decision",
          "content_hash"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
...
      "CloseRequest": {
        "additionalProperties": false,
        "properties": {
          "reason_code": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Reason Code",
            "type": "string"
          }
        },
        "required": [
          "reason_code"
        ],
        "title": "CloseRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HoldView": {
        "properties": {
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "member_id": {
            "format": "uuid",
            "title": "Member Id",
            "type": "string"
          },
          "member_name": {
            "title": "Member Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "member_id",
          "member_name",
          "currency",
          "amount_minor",
          "status"
        ],
        "title": "HoldView",
        "type": "object"
      },
...
      "KillRequest": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "member_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Id"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "scope": {
            "enum": [
              "GLOBAL",
              "MEMBER"
            ],
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "scope",
          "enabled",
          "reason"
        ],
        "title": "KillRequest",
        "type": "object"
      },
      "KillView": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "member_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scope",
          "member_id",
          "enabled",
          "reason"
        ],
        "title": "KillView",
        "type": "object"
      },
...
      "LedgerLine": {
        "description": "One of the member's own ledger movements. The counterparty is always Mesh settlement.",
        "properties": {
          "account": {
            "enum": [
              "BALANCE",
              "HELD",
              "CARRIED"
            ],
            "title": "Account",
            "type": "string"
          },
          "amount": {
            "$ref": "#/components/schemas/Money"
          },
          "counterparty": {
            "default": "Mesh settlement",
            "title": "Counterparty",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "entry_kind": {
            "title": "Entry Kind",
            "type": "string"
          }
        },
        "required": [
          "entry_kind",
          "account",
          "description",
          "amount",
          "created_at"
        ],
        "title": "LedgerLine",
        "type": "object"
      },
...
      "MemberRow": {
        "properties": {
          "funding_blocked": {
            "title": "Funding Blocked",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "state",
          "funding_blocked"
        ],
        "title": "MemberRow",
        "type": "object"
      },
      "MemberSettlement": {
        "description": "The member's own side of settlement (MC-PRV-02): no other member is ever named.",
        "properties": {
          "amount": {
            "$ref": "#/components/schemas/Money"
          },
          "carried": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "counterparty": {
            "default": "Mesh settlement",
            "title": "Counterparty",
            "type": "string"
          },
          "held": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "instruction": {
            "enum": [
              "DEBIT",
              "CREDIT",
              "NONE"
            ],
            "title": "Instruction",
            "type": "string"
          },
          "ledger": {
            "items": {
              "$ref": "#/components/schemas/LedgerLine"
            },
            "title": "Ledger",
            "type": "array"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "settled_by_netting": {
            "title": "Settled By Netting",
            "type": "integer"
          },
          "settled_by_transfer": {
            "title": "Settled By Transfer",
            "type": "integer"
          },
          "status": {
            "enum": [
              "PENDING",
              "FUNDS_HELD",
              "SETTLED",
              "CANCELLED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "instruction",
          "amount",
          "reference",
          "held",
          "carried",
          "settled_by_netting",
          "settled_by_transfer",
          "ledger"
        ],
        "title": "MemberSettlement",
        "type": "object"
      },
...
      "Money": {
        "properties": {
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "amount_minor",
          "currency"
        ],
        "title": "Money",
        "type": "object"
      },
...
      "ReasonRequest": {
        "additionalProperties": false,
        "description": "Every Wise staff action carries a reason code (MC-OPS-01).",
        "properties": {
          "reason_code": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Reason Code",
            "type": "string"
          }
        },
        "required": [
          "reason_code"
        ],
        "title": "ReasonRequest",
        "type": "object"
      },
...
      "RunView": {
        "properties": {
          "approvals": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Approvals",
            "type": "array"
          },
          "computations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Computations",
            "type": "array"
          },
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "current_attempt": {
            "title": "Current Attempt",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "own_positions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Own Positions",
            "type": "array"
          },
          "settlement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemberSettlement"
              },
              {
                "type": "null"
              }
            ]
          },
          "settlement_detail": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SettlementDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "statement_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statement Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "timeline": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Timeline",
            "type": "array"
          }
        },
        "required": [
          "id",
          "status",
          "current_attempt",
          "started_at"
        ],
        "title": "RunView",
        "type": "object"
      },
...
      "SettlementDetail": {
        "description": "Wise staff view of a run's settlement: holds, jobs, transfers and clearing.",
        "properties": {
          "clearing": {
            "items": {
              "$ref": "#/components/schemas/AccountBalance"
            },
            "title": "Clearing",
            "type": "array"
          },
          "commit_entry_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Entry Seq"
          },
          "holds": {
            "items": {
              "$ref": "#/components/schemas/HoldView"
            },
            "title": "Holds",
            "type": "array"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobView"
            },
            "title": "Jobs",
            "type": "array"
          },
          "transfers": {
            "items": {
              "$ref": "#/components/schemas/TransferView"
            },
            "title": "Transfers",
            "type": "array"
          }
        },
        "required": [
          "holds",
          "jobs",
          "transfers",
          "clearing",
          "commit_entry_seq"
        ],
        "title": "SettlementDetail",
        "type": "object"
      },
...
      "StatementView": {
        "properties": {
          "approval_count": {
            "title": "Approval Count",
            "type": "integer"
          },
          "can_approve": {
            "title": "Can Approve",
            "type": "boolean"
          },
          "content": {
            "additionalProperties": true,
            "title": "Content",
            "type": "object"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "current": {
            "title": "Current",
            "type": "boolean"
          },
          "current_statement_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Statement Id"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "required_approvers": {
            "title": "Required Approvers",
            "type": "integer"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "run_status": {
            "title": "Run Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "run_status",
          "current",
          "current_statement_id",
          "content_hash",
          "content",
          "required_approvers",
          "approval_count",
          "can_approve",
          "issued_at",
          "expires_at"
        ],
        "title": "StatementView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Multilateral invoice netting and settlement (hackathon prototype).",
    "title": "Wise Mesh Continental API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/kill-switches": {
      "put": {
        "operationId": "switches_v1_admin_kill_switches_put",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Switches",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/admin/runs/{run_id}": {
      "get": {
        "operationId": "admin_run_v1_admin_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/admin/runs/{run_id}/abort": {
      "post": {
        "description": "Release every hold and return the run's invoices for the next window.",
        "operationId": "abort_v1_admin_runs__run_id__abort_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Abort",
        "tags": [
          "settlement"
        ]
      }
    },
    "/v1/admin/runs/{run_id}/settle": {
      "post": {
        "description": "Prepare and commit; repeating it resumes an interrupted settlement or does nothing.",
        "operationId": "settle_v1_admin_runs__run_id__settle_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Settle",
        "tags": [
          "settlement"
        ]
      }
    },
    "/v1/admin/windows/current/close": {
      "post": {
        "operationId": "close_v1_admin_windows_current_close_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/runs/{run_id}": {
      "get": {
        "operationId": "run_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/statements/{statement_id}": {
      "get": {
        "operationId": "statement_v1_statements__statement_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "statement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Statement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatementView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Statement",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/statements/{statement_id}/approvals": {
      "post": {
        "operationId": "approve_v1_statements__statement_id__approvals_post",
        "parameters": [
          {
            "in": "path",
            "name": "statement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Statement Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}